refactor: address PR review comments on wasm backend
- Split backend/wasm/src/lib.rs into emit.rs (Codegen + Emitter) and
  types.rs (type helpers), keeping lib.rs as the thin entry point
- Replace binop! macro with emit_binop helper method
- Remove U0 early-return guard in emit_prim_app; handle U0 in
  bitnot_mask instead (U0 arithmetic works naturally: 0 op 0 = 0)
- Add cast_signed() comments explaining u64>=2^63 round-trip correctness
- Add u64_large_lit snapshot test exercising values >= 2^63
- Make Phase::Wasm unconditional in snap.rs; only the impl block
  calling splic_backend_wasm is cfg-gated
