+++ driver/tests/snap.rs
@@ -16,15 +16,10 @@ enum Phase {
     Parse,
     Check,
     Stage,
-    #[cfg(feature = "backend-wasm")]
     Wasm,
 }
 
 impl Phase {
-    #[cfg(not(feature = "backend-wasm"))]
-    const ALL: &'static [Self] = &[Self::Lex, Self::Parse, Self::Check, Self::Stage];
-
-    #[cfg(feature = "backend-wasm")]
     const ALL: &'static [Self] = &[Self::Lex, Self::Parse, Self::Check, Self::Stage, Self::Wasm];
 
     const fn snap_filename(self) -> &'static str {
@@ -33,7 +28,6 @@ impl Phase {
             Self::Parse => "2_parse.txt",
             Self::Check => "3_check.txt",
             Self::Stage => "6_stage.txt",
-            #[cfg(feature = "backend-wasm")]
             Self::Wasm => "8_wasm.wat",
         }
     }
@@ -75,13 +69,8 @@ impl ExpectedOutcome {
 ///
 /// When adding a new folder under `tests/snap/`, add a corresponding entry here.
 fn expected_outcome(folder: &str) -> ExpectedOutcome {
-    #[cfg(feature = "backend-wasm")]
-    let stage_success = ExpectedOutcome::run_till(Phase::Wasm);
-    #[cfg(not(feature = "backend-wasm"))]
-    let stage_success = ExpectedOutcome::run_till(Phase::Stage);
-
     match folder {
-        "full" | "examples" => stage_success,
+        "full" | "examples" => ExpectedOutcome::run_till(Phase::Wasm),
         "lex" => ExpectedOutcome::run_till(Phase::Lex),
         "lex_error" => ExpectedOutcome::fail_at(Phase::Lex),
         "parse_error" => ExpectedOutcome::fail_at(Phase::Parse),
@@ -153,6 +142,9 @@ macro_rules! phase {
 ///
 /// The top-level folder name determines which phase (if any) is expected to fail and
 /// which later snapshots must be absent; see `expected_outcome` for the mapping.
+///
+/// The Wasm phase (8) requires the `backend-wasm` feature. When the feature is absent
+/// the step is skipped and the test exits at Stage without writing `8_wasm.wat`.
 #[rstest]
 #[timeout(std::time::Duration::from_secs(if cfg!(miri) { 600 } else { 5 }))]
 fn snap(#[files("tests/snap/*/*/0_input.splic")] path: PathBuf) {
@@ -203,6 +195,8 @@ fn snap(#[files("tests/snap/*/*/0_input.splic")] path: PathBuf) {
 
     // ── Phase 8: Wasm ────────────────────────────────────────────────────────
     // (slot 7 reserved for a future optimisation pass)
+    // When the backend-wasm feature is absent this step is a no-op: no snapshot
+    // is written and the pipeline ends at Stage rather than Wasm.
     #[cfg(feature = "backend-wasm")]
     {
         let wasm_result = splic_backend_wasm::compile_wasm(&staged)
@@ -213,8 +207,6 @@ fn snap(#[files("tests/snap/*/*/0_input.splic")] path: PathBuf) {
         };
         phase!(wasm_snap, wasm_result, Phase::Wasm, expected, dir);
     }
-
-    // Suppress unused-variable warning when backend-wasm is disabled.
     #[cfg(not(feature = "backend-wasm"))]
     let _ = staged;
 }
