+++ driver/tests/snap/full/u64_large_lit/1_lex.txt
@@ -1,0 +1,48 @@
+Code
+Fn
+Ident("min_signed")
+LParen
+RParen
+Arrow
+Ident("u64")
+LBrace
+Num(9223372036854775808)
+RBrace
+Code
+Fn
+Ident("max_u64")
+LParen
+RParen
+Arrow
+Ident("u64")
+LBrace
+Num(18446744073709551615)
+RBrace
+Code
+Fn
+Ident("clamp_large")
+LParen
+Ident("x")
+Colon
+Ident("u64")
+RParen
+Arrow
+Ident("u64")
+LBrace
+Match
+Ident("x")
+LBrace
+Num(9223372036854775808)
+DArrow
+Num(0)
+Comma
+Num(18446744073709551615)
+DArrow
+Num(1)
+Comma
+Ident("n")
+DArrow
+Ident("n")
+Comma
+RBrace
+RBrace
