+++ driver/tests/snap/full/u64_large_lit/2_parse.txt
@@ -1,0 +1,55 @@
+Program {
+    functions: [
+        Function {
+            phase: Object,
+            name: "min_signed",
+            params: [],
+            ret_ty: "u64",
+            body: Block {
+                stmts: [],
+                expr: 9223372036854775808,
+            },
+        },
+        Function {
+            phase: Object,
+            name: "max_u64",
+            params: [],
+            ret_ty: "u64",
+            body: Block {
+                stmts: [],
+                expr: 18446744073709551615,
+            },
+        },
+        Function {
+            phase: Object,
+            name: "clamp_large",
+            params: [
+                Param {
+                    name: "x",
+                    ty: "u64",
+                },
+            ],
+            ret_ty: "u64",
+            body: Block {
+                stmts: [],
+                expr: Match {
+                    scrutinee: "x",
+                    arms: [
+                        MatchArm {
+                            pat: 9223372036854775808,
+                            body: 0,
+                        },
+                        MatchArm {
+                            pat: 18446744073709551615,
+                            body: 1,
+                        },
+                        MatchArm {
+                            pat: "n",
+                            body: "n",
+                        },
+                    ],
+                },
+            },
+        },
+    ],
+}
