+++ driver/tests/snap/full/u64_large_lit/6_stage.txt
@@ -1,0 +1,16 @@
+code fn min_signed() -> u64 {
+    9223372036854775808_u64
+}
+
+code fn max_u64() -> u64 {
+    18446744073709551615_u64
+}
+
+code fn clamp_large(x@0: u64) -> u64 {
+    match x@0 {
+        9223372036854775808 => 0_u64,
+        18446744073709551615 => 1_u64,
+        n@1 => n@1,
+    }
+}
+
